+++ pkgs/yaml_edit/test/update_test.dart
@@ -723,6 +723,71 @@ c: 3
           'YAML': "YAML Ain't Markup Language",
         });
       });
+
+      test('spanning multiple lines with trailing comma', () {
+        final doc = YamlEditor('''
+{
+  a: 1,
+  b: 2,
+}
+''');
+        doc.update(['c'], 3);
+        expect(doc.toString(), equals('''
+{
+  a: 1,
+  b: 2,
+c: 3,}
+'''));
+        expectYamlBuilderValue(doc, {'a': 1, 'b': 2, 'c': 3});
+      });
+
+      test('spanning multiple lines with trailing comma and comment', () {
+        final doc = YamlEditor('''
+{
+  a: 1,
+  b: 2, # comment
+}
+''');
+        doc.update(['c'], 3);
+        expect(doc.toString(), equals('''
+{
+  a: 1,
+  b: 2, # comment
+c: 3,}
+'''));
+        expectYamlBuilderValue(doc, {'a': 1, 'b': 2, 'c': 3});
+      });
+
+      test('spanning multiple lines without trailing comma', () {
+        final doc = YamlEditor('''
+{
+  a: 1,
+  b: 2
+}
+''');
+        doc.update(['c'], 3);
+        expect(doc.toString(), equals('''
+{
+  a: 1,
+  b: 2
+, c: 3}
+'''));
+        expectYamlBuilderValue(doc, {'a': 1, 'b': 2, 'c': 3});
+      });
+
+      test('single line with trailing comma', () {
+        final doc = YamlEditor('{a: 1, b: 2, }');
+        doc.update(['c'], 3);
+        expect(doc.toString(), equals('{a: 1, b: 2, c: 3,}'));
+        expectYamlBuilderValue(doc, {'a': 1, 'b': 2, 'c': 3});
+      });
+
+      test('single line with trailing comma (no space)', () {
+        final doc = YamlEditor('{a: 1, b: 2,}');
+        doc.update(['c'], 3);
+        expect(doc.toString(), equals('{a: 1, b: 2, c: 3,}'));
+        expectYamlBuilderValue(doc, {'a': 1, 'b': 2, 'c': 3});
+      });
     });
 
     group('block map', () {